move web resources, pivot away from feature flag
