+++ study/src/org/labkey/study/importer/AssayScheduleImporter.java
@@ -65,6 +65,9 @@ public class AssayScheduleImporter extends DefaultStudyDesignImporter implements
         if (!ctx.isDataTypeSelected(getDataType()))
             return;
 
+        if (!isStudyDesignEnabled(ctx.getContainer()))
+            return;
+
         if (isValidForImportArchive(ctx, root))
         {
             ExportDirType dirType = ctx.getXml().getAssaySchedule();
