+++ study/src/org/labkey/study/writer/AssayScheduleWriter.java
@@ -55,6 +55,9 @@ public class AssayScheduleWriter extends DefaultStudyDesignWriter implements Int
     @Override
     public void write(StudyImpl object, StudyExportContext ctx, VirtualFile root) throws Exception
     {
+        if (!isStudyDesignEnabled(ctx.getContainer()))
+            return;
+
         StudyDocument.Study studyXml = ctx.getXml();
 
         ExportDirType dir = studyXml.addNewAssaySchedule();
